+++ lib/model/player.dart
@@ -1,0 +1,15 @@
+import 'package:team_draw/model/position.dart';
+
+class Player {
+  String name;
+  Position principalPosition;
+  Position? secondaryPosition;
+  int overall;
+
+  Player({
+    required this.name,
+    required this.overall,
+    required this.principalPosition,
+    this.secondaryPosition,
+  });
+}
